2005-06-13  Jeremy Katz  <katzj@redhat.com>
	* Makefile Makefile.inc mini-wm.c xmouse.c xutils.c isys/Makefile
	isys/isys.c isys/lang.c isys/smp.c isys/uncpio.c
	isys/gzlib/Makefile isys/gzlib/inflate.c loader2/Makefile
	loader2/driverdisk.c loader2/driverselect.c loader2/hdinstall.c
	loader2/init.c loader2/kickstart.c loader2/loader.c
	loader2/loadermisc.c loader2/moduleinfo.h loader2/modules.c
	loader2/net.c loader2/nfsinstall.c loader2/telnet.c
	loader2/undomounts.c loader2/urlinstall.c utils/Makefile
	utils/genhdlist.c utils/mapshdr.c utils/snarffont.c: Define cflags
	in one place (Makefile.inc) and use a consistent set
	everywhere. Fix up some of the compile warnings this uncovered.
	Also fix a few more things that -Wextra turned up.

+++ isys/uncpio.c
@@ -79,7 +79,7 @@ struct cpioHeader {
     gid_t gid;
     int nlink;
     time_t mtime;
-    long size;
+    unsigned long size;
     dev_t dev, rdev;
     char * path;
 };
@@ -304,7 +304,7 @@ static int expandRegular(struct ourfd * fd, struct cpioHeader * hdr,
     int out;
     char buf[8192];
     int bytesRead;
-    int left = hdr->size;
+    unsigned long left = hdr->size;
     int rc = 0;
     struct cpioCallbackInfo cbInfo;
     struct stat sb;
@@ -449,9 +449,9 @@ static int createLinks(struct hardLink * li, const char ** failedFile) {
     return 0;
 }
 
-static int eatBytes(struct ourfd * fd, int amount) {
+static int eatBytes(struct ourfd * fd, unsigned long amount) {
     char buf[4096];
-    int bite;
+    unsigned long bite;
    
     while (amount) {
 	bite = (amount > sizeof(buf)) ? sizeof(buf) : amount;
